Several changes: use the more efficient getEntrySet instead of keySet then getting values. Avoid work and memory allocations that were completely unusued. Make a number of fields static (saves memory per instance). Move some unused classes to the archive section.
