+++ WEB-INF/classes/dods/dap/parser/DASParserTokenManager.java
@@ -726,7 +726,7 @@ public class DASParserTokenManager implements DASParserConstants {
   }
 
   int curLexState = 0;
-  final int defaultLexState = 0;
+  static final int defaultLexState = 0;
   int jjnewStateCnt;
   int jjround;
   int jjmatchedPos;
