+++ WEB-INF/classes/dods/dap/parser/DDSParser.java
@@ -716,7 +716,6 @@ public class DDSParser implements DDSParserConstants {
   private int jj_ntk;
   private Token jj_scanpos, jj_lastpos;
   private int jj_la;
-  public final boolean lookingAhead = false;
   private int jj_gen;
   private final int[] jj_la1 = new int[22];
   private final int[] jj_la1_0 = {
@@ -842,7 +841,7 @@ public class DDSParser implements DDSParserConstants {
   }
 
   public final Token getToken(int index) {
-    Token t = lookingAhead ? jj_scanpos : token;
+    Token t = token;
     for (int i = 0; i < index; i++) {
       if (t.next != null) t = t.next;
       else t = t.next = token_source.getNextToken();
