+++ WEB-INF/classes/dods/dap/parser/ErrorParserTokenManager.java
@@ -583,7 +583,7 @@ public class ErrorParserTokenManager implements ErrorParserConstants {
   }
 
   int curLexState = 0;
-  final int defaultLexState = 0;
+  static final int defaultLexState = 0;
   int jjnewStateCnt;
   int jjround;
   int jjmatchedPos;
