+++ WEB-INF/classes/gov/noaa/pfel/erddap/dataset/EDDTable.java
@@ -20824,78 +20824,6 @@ public abstract class EDDTable extends EDD {
   */
 
   /**
-   * This generates an OPeNDAP query from a Google Visualization API Query Language string. <br>
-   * See Query Language reference
-   * https://developers.google.com/chart/interactive/docs/querylanguage?csw=1
-   *
-   * @param gVisQuery the case of the key words (e.g., SELECT) is irrelevant
-   * @return dapQuery
-   * @throws Throwable if trouble ???or if error, return gVis-style error message string?
-   */
-  public String gVisQueryToDapQuery(String gVisQuery) {
-    // SELECT [DISTINCT] csvRequestColumns [WHERE constraint [AND constraint]] [ORDER BY csvOrderBy]
-    // csvRequestColumns[&constraint][&distinct()][&orderBy csvOrderBy
-    int gVisQueryLength = gVisQuery.length();
-    String gVisQueryLC = gVisQuery.toLowerCase();
-    StringBuilder dapQuery = new StringBuilder();
-
-    // SELECT
-    int start = 0;
-    if (gVisQueryLC.startsWith("select ", start)) {
-      start += 7;
-      while (gVisQueryLC.startsWith(" ", start)) start++;
-      if (gVisQueryLC.startsWith("*", start)) {
-        start++;
-        while (gVisQueryLC.startsWith(" ", start)) start++;
-      } else {
-        do {
-          if (dapQuery.length() > 0) {
-            // eat the required ',' and optional white space
-            start++;
-            while (gVisQueryLC.startsWith(" ", start)) start++;
-            // add , to dapQuery
-            dapQuery.append(',');
-          }
-          String name;
-          if (gVisQueryLC.startsWith("'", start)) {
-            // 'colName'
-            int po = gVisQueryLC.indexOf("'", start + 1);
-            if (po < 0)
-              return EDStatic.messages.queryErrorAr[0] + "missing \"'\" after position=" + start;
-            name = gVisQuery.substring(start + 1, po);
-            start = po + 1;
-          } else {
-            // unquoted colName
-            int po = start + 1;
-            while (po < gVisQueryLength
-                && !gVisQueryLC.startsWith(",", po)
-                && !gVisQueryLC.startsWith(" ", po)) po++;
-            name = gVisQuery.substring(start, po);
-            start = po;
-          }
-          if (String2.indexOf(dataVariableDestinationNames, name) < 0)
-            return EDStatic.messages.queryErrorAr[0]
-                + "unrecognized column name=\""
-                + name
-                + "\" at position="
-                + (start - name.length());
-          dapQuery.append(name);
-          while (gVisQueryLC.startsWith(" ", start)) start++;
-
-        } while (gVisQueryLC.startsWith(",", start));
-      }
-    }
-
-    // WHERE
-    if (gVisQueryLC.startsWith("where ", start)) {}
-
-    // ORDER BY
-    if (gVisQueryLC.startsWith("order by ", start)) {}
-
-    return dapQuery.toString();
-  }
-
-  /**
    * This writes the dataset's FGDC-STD-001-1998 "Content Standard for Digital Geospatial Metadata"
    * https://www.fgdc.gov/standards/projects/FGDC-standards-projects/metadata/base-metadata/index_html
    * (unlike EDDGrid, not "Extensions for Remote Sensing Metadata") XML to the writer. <br>
