+++ WEB-INF/classes/gov/noaa/pfel/erddap/util/EDMessages.java
@@ -1074,7 +1074,7 @@ public class EDMessages {
   public final int[] pdfWidths;
   public final int[] pdfHeights;
   private final String[] theLongDescriptionHtmlAr; // see the xxx() methods
-  public final String errorFromDataSource = String2.ERROR + " from data source: ";
+  public static final String errorFromDataSource = String2.ERROR + " from data source: ";
   public final int nLanguages = TranslateMessages.languageList.size();
 
   public EDMessages(String contentDirectory) throws Exception {
