+++ WEB-INF/classes/gov/noaa/pmel/sgt/MonthYearAxis.java
@@ -36,7 +36,7 @@ public class MonthYearAxis implements TimeAxisStyle {
   static final String defaultMajorLabelFormat__ = "yyy";
   static final int defaultNumSmallTics__ = 0;
   int defaultMinorLabelInterval_ = 2;
-  final int defaultMajorLabelInterval_ = 1;
+  static final int defaultMajorLabelInterval_ = 1;
   static final double incrementValue__ = 1.0;
   static final int incrementUnits__ = GeoDate.MONTHS;
