+++ pom.xml
@@ -412,6 +412,7 @@
                 <version>4.9.3.0</version>
                 <configuration>
                     <effort>max</effort>
+                    <maxHeap>2048</maxHeap>
                     <includeFilterFile>spotbugs-security-include.xml</includeFilterFile>
                 </configuration>
             </plugin>
